+++ src/test/groovy/nextflow/slack/SlackConfigTest.groovy
@@ -506,7 +506,6 @@ class SlackConfigTest extends Specification {
                 seqeraPlatform: [
                     enabled: true,
                     actionButtons: [
-                        mode: 'interactive',
                         cancel: false,
                         resume: true,
                         relaunch: false
@@ -519,10 +518,8 @@ class SlackConfigTest extends Specification {
         def config = SlackConfig.from(session)
 
         then:
-        config.seqeraPlatform.actionButtons.mode == 'interactive'
         !config.seqeraPlatform.actionButtons.cancel
         config.seqeraPlatform.actionButtons.resume
         !config.seqeraPlatform.actionButtons.relaunch
-        config.seqeraPlatform.actionButtons.isInteractiveMode()
     }
 }